move event table into components
