+++ frontend/components/tables/event-table/event-table.tsx
